add a very basic executable
The idea was that this would be a quick simple thing with basically
no UI, which just read in lines until STDIN closed, and formatted
each one.

However, it has an error which doesn't seem to make much sense:

```
error[E0597]: `line` does not live long enough
  --> src/main.rs:19:20
   |
19 |         let line = line.trim();
   |                    ^^^^-------
   |                    |
   |                    borrowed value does not live long enough
   |                    argument requires that `line` is borrowed for `'static`
...
22 |     }
   |     - `line` dropped here while still borrowed
```

It is very much non-obvious what argument is imposing the 'static
requirement, or how to get rid of it.